+++ deploy/docker-compose.yml
@@ -66,7 +66,7 @@ services:
     container_name: laravel-shop-nginx
     restart: unless-stopped
     ports:
-      - "80:80"
+      - "8000:8000"
     volumes:
       - ../project:/var/www/html
       - ./nginx.conf:/etc/nginx/nginx.conf
