+++ deploy/nginx.conf
@@ -19,7 +19,7 @@ http {
     keepalive_timeout  65;
 
     server {
-        listen 80;
+        listen 8000;
         index index.php index.html;
         root /var/www/html/public;
