+++ project/app/Http/Controllers/Api/ProductController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\StoreProductRequest;
 use App\Http\Requests\Api\UpdateProductRequest;
 use App\Models\Product;
+use App\Services\ProductService;
 use Illuminate\Http\Request;
 
 class ProductController extends Controller
@@ -13,67 +14,50 @@ class ProductController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(ProductService $productService)
     {
-        return Product::all();
+        return $productService->index();
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreProductRequest $request)
+    public function store(ProductService $productService, StoreProductRequest $request, Product $product)
     {
-        if (!AuthController::isAdmin($request)) {
+        if (auth()->user()->can('create', $product)) {
+            return $productService->store($request);
+        } else {
             return response()->json([
                 'message' => 'Forbidden for you'
             ], 403);
         }
-        Product::query()->create($request->all());
-        $id = Product::query()->get()->last()->id;
-        return response()->json([
-            'id' => $id,
-            'message' => 'Product added'
-        ], 201);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateProductRequest $request, $id)
+    public function update(ProductService $productService, UpdateProductRequest $request, Product $product)
     {
-        if (!AuthController::isAdmin($request)) {
+        if (auth()->user()->can('update', $product)) {
+            return $productService->update($request);
+        } else {
             return response()->json([
                 'message' => 'Forbidden for you'
             ], 403);
         }
-        if (preg_match('/^[0-9]+$/', $id) && Product::where('id', $id)->exists()) {
-            $product = Product::query()->find($id);
-            $product->update($request->all());
-            return response()->json([$product]);
-        }
-        return response()->json([
-            'message' => 'Not found'
-        ], 404);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Request $request, $id)
+    public function destroy(ProductService $productService, Request $request, Product $product)
     {
-        if (!AuthController::isAdmin($request)) {
+        if (auth()->user()->can('delete', $product)) {
+            return $productService->destroy($request);
+        } else {
             return response()->json([
                 'message' => 'Forbidden for you'
             ], 403);
         }
-        if (preg_match('/^[0-9]+$/', $id) && Product::where('id', $id)->exists()) {
-            Product::where('id', $id)->delete();
-            return response()->json([
-                'message' => 'Product removed'
-            ]);
-        }
-        return response()->json([
-            'message' => 'Not found'
-        ], 404);
     }
 }
