+++ project/app/Policies/CartItemPolicy.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Policies;
+
+use App\Models\CartItem;
+use App\Models\User;
+
+class CartItemPolicy
+{
+
+    /**
+     * Determine whether the user can view the model.
+     */
+    public function view(User $user, CartItem $cartItem): bool
+    {
+        return !$user->is_admin;
+    }
+
+    /**
+     * Determine whether the user can create models.
+     */
+    public function create(User $user, CartItem $cartItem): bool
+    {
+        return !$user->is_admin;
+    }
+
+    /**
+     * Determine whether the user can delete the model.
+     */
+    public function delete(User $user, CartItem $cartItem): bool
+    {
+        return !$user->is_admin;
+    }
+
+}
