+++ project/app/Policies/ProductPolicy.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Policies;
+
+use App\Models\Product;
+use App\Models\User;
+
+class ProductPolicy
+{
+
+    /**
+     * Determine whether the user can create models.
+     */
+    public function create(User $user, Product $product): bool
+    {
+        return (bool)$user->is_admin;
+    }
+
+    /**
+     * Determine whether the user can update the model.
+     */
+    public function update(User $user, Product $product): bool
+    {
+        return (bool)$user->is_admin;
+    }
+
+    /**
+     * Determine whether the user can delete the model.
+     */
+    public function delete(User $user, Product $product): bool
+    {
+        return (bool)$user->is_admin;
+    }
+
+}
